--- js/components/form/format-completer.js
@@ -1,24 +1,0 @@
-define([
-    'vue',
-    'api',
-    'components/form/base-completer.vue'
-], function(Vue, API, BaseCompleter) {
-    'use strict';
-
-    return {
-        mixins: [BaseCompleter],
-        ns: 'datasets',
-        endpoint: 'suggest_formats',
-        selectize: {
-            maxItems: 1,
-            valueField: 'text',
-            plugins: ['remove_button'],
-            create: function(input) {
-                return {
-                    value: input,
-                    text: input
-                }
-            }
-        }
-    };
-});
