+++ js/components/form/format-completer.vue
@@ -1,0 +1,23 @@
+<!-- Formats autocompleter -->
+<script>
+import Vue from 'vue';
+import API from 'api';
+import BaseCompleter from 'components/form/base-completer.vue';
+
+export default {
+    mixins: [BaseCompleter],
+    ns: 'datasets',
+    endpoint: 'suggest_formats',
+    selectize: {
+        maxItems: 1,
+        valueField: 'text',
+        plugins: ['remove_button'],
+        create: function(input) {
+            return {
+                value: input,
+                text: input
+            }
+        }
+    }
+};
+</script>
