--- js/components/form/image-field.vue
@@ -1,156 +1,0 @@
-<style lang="less">
-.image-picker-field {
-    .image-picker-preview {
-        .pull-left();
-        width: 100px;
-        height: 100px;
-
-        margin-right: 10px;
-
-        img {
-            max-width: 100%;
-            max-height: 100%;
-        }
-    }
-}
-</style>
-
-<template>
-<div class="image-picker-field" data-sizes="{{sizes}}" data-basename="{{field.name}}" data-endpoint="{{field.endpoint}}">
-    <div class="image-picker-preview">
-        <img v-el="preview" src="{{field.src || default_img}}" data-placeholder="{{default_img}}"/>
-    </div>
-    <span class="image-picker-btn btn btn-default btn-file"
-        v-el="btn" v-on="click: click">
-        {{ _('Choose a picture') }}
-            <!-- field.filename(),
-
-            field.bbox(), -->
-    '</span>',
-</div>
-</template>
-
-<script>
-'use strict';
-
-var HAS_FILE_API = window.File && window.FileReader && window.FileList && window.Blob,
-    log = require('logger');
-
-module.exports = {
-    data: function() {
-        return {};
-    },
-    // ready: function(el, opts) {
-    //     // this.$el = $(el);
-    //     // this.opts = opts;
-    //     // this.basename = this.$el.data('basename');
-
-    //     // this.$btn = this.$el.find('.image-picker-btn');
-    //     // this.$file = this.$el.find('input:file');
-    //     // this.$bbox = this.$el.find('#' + this.basename + '-bbox');
-    //     // this.$filename = this.$el.find('#' + this.basename + '-filename');
-    //     // this.$preview = this.$el.find('.image-picker-preview img');
-
-    //     this.$btn.click(this.f('on_click'));
-    //     this.endpoint = this.$el.data('endpoint');
-    //     this.sizes = $((this.$el.data('sizes') || '100').toString().split(','))
-    //         .map(function(idx, value) { return parseInt(value); })
-    //         .filter(function(idx, value) { return !isNaN(value); })
-    //         .get();
-
-    //     log.info('ImagePicker initialized', this.sizes);
-    // },
-    methods: {
-        on_click: function() {
-            this.$modal = modal({
-                title: i18n._('Choose an image'),
-                content: tpl(),
-                size: 'lg',
-                close_btn: i18n._('Cancel'),
-                close_cls: 'btn-warning',
-                actions: [{
-                    label: i18n._('Submit'),
-                    icon: 'fa-check',
-                    classes: 'btn-info btn-submit hide'
-                }, {
-                    label: i18n._('Clear'),
-                    icon: 'fa-trash',
-                    classes: 'btn-info btn-danger btn-clear pull-right hide'
-                }]
-            });
-
-            this.$btn_submit = this.$modal.find('.btn-submit');
-            this.$btn_clear = this.$modal.find('.btn-clear');
-
-            this.uploader = new Uploader(this.$modal.find('.uploader'), {
-                endpoint: this.endpoint,
-                auto: false
-            });
-            this.$cropper = this.$modal.find('.image-cropper-container')
-
-            $(this.uploader)
-                .on('file-picked', this.f('on_file_picked'))
-                .on('complete', this.f('on_upload_complete'));
-
-            this.$btn_clear.click(this.f('clear'));
-            this.$btn_submit.click(this.f('on_submit'));
-
-            return false;
-        },
-
-        crop: function(image) {
-            this.$modal.find('section').addClass('hide');
-            this.$btn_submit.removeClass('hide');
-            this.$btn_clear.removeClass('hide');
-            this.$modal.find('.modal-title').text(i18n._('Crop your thumbnail'));
-            this.$cropper.removeClass('hide');
-
-            this.cropper = new Cropper(this.$cropper, {
-                sizes: this.sizes
-            });
-
-            this.cropper.load(image);
-        },
-
-        clear: function() {
-            this.$modal.find('.image-cropper-container').addClass('hide');
-            this.$modal.find('.uploader').removeClass('hide');
-            this.uploader.clear();
-        },
-
-        on_file_picked: function(ev, file, name) {
-            if (HAS_FILE_API) {
-                this.crop(URL.createObjectURL(file));
-            } else {
-                log.warning('File APIs not supported');
-                this.uploader.upload();
-            }
-        },
-
-        on_upload_complete: function(ev, name, response) {
-            this.filename = response.filename;
-            this.tmp_url = response.url;
-            if (HAS_FILE_API) {
-                this.save();
-            } else {
-                this.crop(response.url);
-            }
-        },
-
-        on_submit: function() {
-            if (HAS_FILE_API) {
-                this.uploader.upload();
-            } else {
-                this.save()
-            }
-        },
-
-        save: function() {
-            this.$modal.modal('hide');
-            this.$bbox.val(this.cropper.get_bbox());
-            this.$filename.val(this.filename);
-            this.$preview.attr('src', this.tmp_url);
-        }
-    }
-};
-</script>
