+++ js/components/form/user-completer.vue
@@ -1,0 +1,40 @@
+<!-- User autocompleter -->
+<script>
+import Vue from 'vue';
+import API from 'api';
+import BaseCompleter from 'components/form/base-completer.vue';
+import placeholders from 'helpers/placeholders';
+
+const template = `<div class="selectize-option">
+    <div class="logo pull-left">
+        <img src="{{user.avatar || user.avatar_url || placeholder}}"/>
+    </div>
+    {{user.fullname}}
+</div>`;
+
+
+function render(user) {
+    return new Vue({data: {user: user, placeholder: placeholders.user}}).$interpolate(template);
+}
+
+export default {
+    name: 'user_completer',
+    mixins: [BaseCompleter],
+    ns: 'users',
+    endpoint: 'suggest_users',
+    selectize: {
+        valueField: 'id',
+        searchField: ['fullname'],
+        options: [],
+        plugins: ['remove_button'],
+        render: {
+            option: function(data, escape) {
+                return render(data);
+            },
+            item: function(data, escape) {
+                return render(data);
+            }
+        },
+    }
+};
+</script>
